backend/requirements.txt
accelerate==1.8.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.55.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argcomplete==3.6.2
async-timeout==5.0.1
attrs==25.3.0
Automat==20.2.0
Babel==2.8.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.13.4
blinker==1.4
boto3==1.38.43
botocore==1.38.43
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
cloud-init==25.1.2
cohere==5.15.0
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
dbus-python==1.2.18
Deprecated==1.2.18
deprecation==2.1.0
dill==0.4.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
docling==2.38.0
docling-core==2.38.2
docling-ibm-models==3.6.0
docling-parse==4.1.0
easyocr==1.7.2
effdet==0.4.1
email_validator==2.2.0
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fasta2a==0.3.3
fastapi==0.115.13
fastapi-cli==0.0.7
fastavro==1.11.1
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
google-genai==1.21.1
googleapis-common-protos==1.70.0
gotrue==2.12.2
griffe==1.7.3
groq==0.28.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.4
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.3
imageio==2.37.0
importlib_metadata==8.7.0
incremental==21.3.0
iniconfig==2.1.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==2.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keyring==23.5.0
kiwisolver==1.4.8
langchain-core==0.3.66
langchain-google-genai==2.1.5
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.2
latex2mathml==3.78.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
logfire-api==3.21.1
lxml==5.4.0
Markdown==3.8.2
markdown-it-py==3.0.0
marko==2.1.4
MarkupSafe==2.0.1
marshmallow==3.26.1
matplotlib==3.10.3
mcp==1.9.4
mdurl==0.1.2
mistralai==1.8.2
more-itertools==8.10.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.5.1
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.0
olefile==0.47
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
openai==1.91.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.34.1
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==0.22.0
pikepdf==9.9.0
pillow==11.2.1
pluggy==1.6.0
postgrest==1.1.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycparser==2.22
pycurl==7.44.1
pydantic==2.11.7
pydantic-ai==0.3.3
pydantic-ai-slim==0.3.3
pydantic-evals==0.3.3
pydantic-graph==0.3.3
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.10.1
pylatexenc==2.10
pyOpenSSL==21.0.0
pypandoc==1.15
pyparsing==2.4.7
pypdf==5.6.0
pypdfium2==4.30.1
pyrsistent==0.18.1
pyserial==3.5
pytest==8.4.1
pytest-mock==3.14.1
python-apt==2.4.0+ubuntu4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2022.1
PyYAML==6.0.2
RapidFuzz==3.13.0
rapidocr-onnxruntime==1.4.4
realtime==2.5.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
rsa==4.9.1
rtree==1.4.0
s3transfer==0.13.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
SecretStorage==3.3.1
semchunk==2.2.2
service-identity==18.1.0
shapely==2.1.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.6
starlette==0.46.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
sympy==1.14.0
systemd-python==234
tabulate==0.9.0
tenacity==8.5.0
tifffile==2025.5.10
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
triton==3.3.1
Twisted==22.1.0
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unstructured==0.17.2
unstructured-client==0.36.0
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
xlrd==2.0.2
xlsxwriter==3.2.5
yarl==1.20.1
zipp==3.23.0
zope.interface==5.4.0
zstandard==0.23.0
